modification of page structure
