--- src/components/Todo.vue
@@ -1,41 +1,0 @@
-<script setup>
-import { onMounted, onUpdated, ref } from "vue";
-
-defineProps({
-  msg: String,
-});
-
-const todos = ref([{ message: "Asdfasdf" }]);
-const newtext = ref("");
-
-function addTodo() {
-  todos.value.push({ message: newtext.value });
-  newtext.value = "";
-}
-
-function removeTodo(index) {
-  todos.value = [...todos.value.filter((todo, id) => id !== index)];
-}
-</script>
-
-<template>
-  <h1>{{ msg }}</h1>
-  <div>
-    <p>Message is: {{ newtext }}</p>
-    <input v-model="newtext" placeholder="edit me" />
-    <button @click="addTodo">Add</button>
-  </div>
-  <div>
-    <li v-for="({ message }, index) in todos">
-      <span>{{ index }}</span>
-      {{ message }}
-      <button @click="removeTodo(index)">X</button>
-    </li>
-  </div>
-</template>
-
-<style scoped>
-li {
-  list-style: none;
-}
-</style>
