+++ src/layout/header.vue
@@ -2,10 +2,10 @@
 <template>
   <header class="header">
     <div>
-      <a href="https://vitejs.dev" target="_blank">
+      <a href="/">
         <img src="../assets/vite.svg" class="logo" alt="Vite logo" />
       </a>
-      <a href="https://vuejs.org/" target="_blank">
+      <a href="/">
         <img src="../assets/vue.svg" class="logo vue" alt="Vue logo" />
       </a>
     </div>
