+++ src/router/index.js
@@ -1,10 +1,24 @@
 import { createMemoryHistory, createRouter } from 'vue-router'
-import HelloWorld from '../components/HelloWorld.vue'
-import Todo from '../components/todo.vue'
+import Home from '../pages/Home.vue'
+import SinglePageTodoList from '../pages/SinglePageTodoList.vue'
+import About from '../pages/about.vue'
 
 const routes = [
-  { path: '/', component: HelloWorld },
-  { path: '/todo', component: Todo },
+  {
+    path: '/',
+    name: Home,
+    component: Home
+  },
+  {
+    path: '/about',
+    name: About,
+    component: About
+  },
+  {
+    path: '/todo',
+    name: SinglePageTodoList,
+    component: SinglePageTodoList
+  },
 ]
 
 const router = createRouter({
