+++ src/style.css
@@ -54,6 +54,8 @@ li {
 h1 {
   font-size: 3.2em;
   line-height: 1.1;
+  text-align: center;
+  margin: 1rem 0;
 }
 
 button {
